+++ models/medicos.py
@@ -11,7 +11,7 @@ class medicos(db.Model):
             'afiliacao hospitalar':self.afiliacaohospitalar
         }
     
-    codigo = db.Column(db.Integer(5), primary_key=True)
+    codigo = db.Column(db.Integer, primary_key=True)
     codusuario = db.Column(ForeignKey('usuarios.codigo'))
     nome = db.Column(db.String(50))
     especializacao = db.Column(db.String(50))
